// src/app/api/interfaces/studentdashboard/MentorInterFace.ts
export interface MentorDetailsProps {
    Mentoruid: number | string;
    dept_name: string;
    MentorName: string;
    School: string;
    Picture: string;
    PreferenceOne: string;
    PreferenceTwo: string;
    PreferenceThree: string;
    headline?: string;
    CheakAcess?: string;
    SeatLeft: number | string;
}